+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ * Added support for auto-resolving namespaces for keyword from the current namespace using the `::kw` syntax (#576)
 
 ## [v0.1.dev14] - 2020-06-18
 ### Added